Rename single/double buffer namespaces
